Rename UserInterface file for better readability
#84vnnm
#84vpdu
