+++ auth_provider/lib/src/AuthUser.dart
